:bug: Fix: enums 분리 및 온보딩 오류 수정
