+++ src/main/java/com/example/chaeum/chaeum_be/enums/PurposeType.java
@@ -1,0 +1,9 @@
+package com.example.chaeum.chaeum_be.enums;
+
+import lombok.Getter;
+
+@Getter
+public enum PurposeType {
+    BUY, SELL, BOTH
+}
+
